add .env stripe still broken

+++ checkout/views.py
@@ -1,19 +1,43 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.conf import settings
+from django.views.decorators.http import require_POST
+from django.http import HttpResponse
+from django.http import JsonResponse
+
+import stripe
+import json
 
 from products.models import Product
 from .models import Order, OrderLineItem
 from .forms import OrderForm
 from bag.contexts import bag_contents
 
-import stripe
+
+@require_POST
+def cache_checkout_data(request):
+    try:
+        pid = request.POST.get('client_secret').split('_secret')[0]
+        stripe.PaymentIntent.modify(
+            pid,
+            metadata={
+                'bag': request.session.get('bag', {}),
+                'save_info': request.POST.get('save_info'),
+                'username': request.user.username if request.user.is_authenticated else 'Anonymous',
+            }
+        )
+        return JsonResponse({'message': 'Success'}, status=200)
+    except Exception as e:
+        return JsonResponse({'error': str(e)}, status=400)
+
 
 def checkout(request):
     stripe_public_key = settings.STRIPE_PUBLIC_KEY
     stripe_secret_key = settings.STRIPE_SECRET_KEY
 
     if request.method == 'POST':
+        print("POST data received:", request.POST)  # Debugging line
+
         bag = request.session.get('bag', {})
 
         form_data = {
@@ -30,39 +54,43 @@ def checkout(request):
         order_form = OrderForm(form_data)
         if order_form.is_valid():
             order = order_form.save(commit=False)
-            pid = request.POST.get('client_secret').split('_secret')[0]
-            order.stripe_pid = pid
-            order.original_bag = json.dumps(bag)
-            order.save()
-            for item_id, item_data in bag.items():
-                try:
-                    product = Product.objects.get(id=item_id)
-                    if isinstance(item_data, int):
-                        order_line_item = OrderLineItem(
-                            order=order,
-                            product=product,
-                            quantity=item_data,
-                        )
-                        order_line_item.save()
-                    else:
-                        for size, quantity in item_data['items_by_size'].items():
+            client_secret = request.POST.get('client_secret')
+            if client_secret:
+                pid = client_secret.split('_secret')[0]
+                order.stripe_pid = pid
+                order.original_bag = json.dumps(bag)
+                order.save()
+                for item_id, item_data in bag.items():
+                    try:
+                        product = Product.objects.get(id=item_id)
+                        if isinstance(item_data, int):
                             order_line_item = OrderLineItem(
                                 order=order,
                                 product=product,
-                                quantity=quantity,
-                                product_size=size,
+                                quantity=item_data,
                             )
                             order_line_item.save()
-                except Product.DoesNotExist:
-                    messages.error(request, (
-                        "One of the products in your bag wasn't found in our database. "
-                        "Please call us for assistance!")
-                    )
-                    order.delete()
-                    return redirect(reverse('view_bag'))
+                        else:
+                            for size, quantity in item_data['items_by_size'].items():
+                                order_line_item = OrderLineItem(
+                                    order=order,
+                                    product=product,
+                                    quantity=quantity,
+                                    product_size=size,
+                                )
+                                order_line_item.save()
+                    except Product.DoesNotExist:
+                        messages.error(request, (
+                            "One of the products in your bag wasn't found in our database. "
+                            "Please call us for assistance!")
+                        )
+                        order.delete()
+                        return redirect(reverse('view_bag'))
 
-            request.session['save_info'] = 'save-info' in request.POST
-            return redirect(reverse('checkout_success', args=[order.order_number]))
+                request.session['save_info'] = 'save-info' in request.POST
+                return redirect(reverse('checkout_success', args=[order.order_number]))
+            else:
+                messages.error(request, 'Missing client_secret.')
         else:
             messages.error(request, 'There was an error with your form. \
                 Please double check your information.')
@@ -96,15 +124,12 @@ def checkout(request):
 
     return render(request, template, context)
 
+
+
 def checkout_success(request, order_number):
-    """
-    Handle successful checkouts
-    """
     save_info = request.session.get('save_info')
     order = get_object_or_404(Order, order_number=order_number)
-    messages.success(request, f'Order successfully processed! \
-        Your order number is {order_number}. A confirmation \
-        email will be sent to {order.email}.')
+    messages.success(request, f'Order successfully processed! Your order number is {order_number}. A confirmation email will be sent to {order.email}.')
 
     if 'bag' in request.session:
         del request.session['bag']
@@ -115,4 +140,3 @@ def checkout_success(request, order_number):
     }
 
     return render(request, template, context)
-
